utils/requirements.txt
boto3==1.9.225
botocore==1.12.225
certifi==2019.6.16
chardet==3.0.4
docutils==0.15.2
ebaysdk==2.1.5
get==2019.4.13
idna==2.8
jmespath==0.9.4
lxml==4.4.1
numpy==1.17.2
pandas==0.25.1
post==2019.4.13
psycopg2-binary==2.8.3
public==2019.4.13
python-dateutil==2.8.0
pytz==2019.2
query-string==2019.4.13
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
